fix for tenantId keycloak type

+++ src/main/java/com/vi/tenantservice/config/security/AuthorisationService.java
@@ -15,8 +15,8 @@ public class AuthorisationService {
     }
 
     public Optional<Long> findCustomUserAttributeInAccessToken(String attributeName) {
-        Long tenantId = (Long) getPrincipal().getKeycloakSecurityContext().getToken().getOtherClaims().get(attributeName);
-        return Optional.ofNullable(tenantId);
+        Integer tenantId = (Integer) getPrincipal().getKeycloakSecurityContext().getToken().getOtherClaims().get(attributeName);
+        return Optional.ofNullable(Long.valueOf(tenantId));
     }
 
     public Object getUsername() {
